refactor http codes to badges
