+++ src/lib/request-logs.ts
@@ -1,32 +1,5 @@
 import { type RequestLogAttributes } from "@/types/request-logs"
 
-export function requestLogMethodBadgeVariant(
-  method: string,
-): "secondary" | "success" | "warning" | "destructive" {
-  switch (method.toUpperCase()) {
-    case "POST":
-      return "success"
-    case "PUT":
-    case "PATCH":
-      return "warning"
-    case "DELETE":
-      return "destructive"
-    default:
-      return "secondary"
-  }
-}
-
-export function requestLogStatusBadgeVariant(
-  status: string,
-): "secondary" | "success" | "warning" | "destructive" {
-  const code = Number(status)
-  if (Number.isNaN(code)) return "secondary"
-  if (code >= 500) return "destructive"
-  if (code >= 400) return "warning"
-  if (code >= 200 && code < 300) return "success"
-  return "secondary"
-}
-
 // attempts to parse string as JSON, but falls back to original
 // string if parsing fails or if value is falsy
 function parseBody(value: string | null | undefined): unknown {
