+++ src/pages/app/request-log/details.tsx
@@ -32,9 +32,8 @@ import { useBackNavigate } from "@/hooks/use-back-navigate"
 import {
   formatRequestLogRequest,
   formatRequestLogResponse,
-  requestLogStatusBadgeVariant,
-  requestLogMethodBadgeVariant,
 } from "@/lib/request-logs"
+import { httpMethodBadgeVariant, httpStatusBadgeVariant } from "@/lib/http"
 import { copyToClipboard } from "@/lib/clipboard"
 
 import { HTTPStatusCodeDescriptions, type HTTPStatusCode } from "@/types/http"
@@ -107,7 +106,7 @@ export default function RequestLogDetails() {
                 <div className="flex flex-wrap items-center gap-2">
                   <TooltipBadge
                     value={requestLog.attributes.method}
-                    variant={requestLogMethodBadgeVariant(
+                    variant={httpMethodBadgeVariant(
                       requestLog.attributes.method,
                     )}
                     tooltip="The HTTP method of the request."
@@ -115,7 +114,7 @@ export default function RequestLogDetails() {
                   />
                   <TooltipBadge
                     value={requestLog.attributes.status}
-                    variant={requestLogStatusBadgeVariant(
+                    variant={httpStatusBadgeVariant(
                       requestLog.attributes.status,
                     )}
                     tooltip={
